demucs/requirements.txt
# please make sure you have already a pytorch install that is cuda enabled!
dora-search>=0.1.12
diffq>=0.2.1
einops
flake8
hydra-colorlog>=1.1
hydra-core>=1.1
julius>=0.2.3
lameenc>=1.2
museval
mypy
openunmix
pyyaml
submitit
torch>=1.8.1
torchaudio>=0.8,<2.3.1
tqdm
treetable
soundfile>=0.10.3
